requirements.txt

faiss-cpu==1.11.0
gradio==5.32.0
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.63
langchain_openai==0.3.18
pypdf==5.6.0
pytest==8.3.5
python-dotenv==1.1.0
Requests==2.32.3
sentence-transformers==4.1.0
